requirements.txt
# Core deep learning frameworks
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.25.0

# Data processing and scientific computing
numpy>=1.24.0
pandas>=2.0.0
scipy>=1.10.0
scikit-learn>=1.3.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
streamlit>=1.25.0

# Time series processing
tsfresh>=0.20.0

# Network and concurrent processing
aiohttp>=3.8.0
websockets>=11.0.0

# Configuration management
pyyaml>=6.0
python-dotenv>=1.0.0

# Logging and monitoring
loguru>=0.7.0

# Data storage
h5py>=3.8.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

# Development tools
jupyter>=1.0.0
ipykernel>=6.25.0
black>=23.0.0

# Additional dependencies
psutil>=5.9.0